calhackproj/python-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
plotly==5.17.0
pylsl==1.16.2
numpy>=1.26.0
scipy>=1.11.2
python-dotenv
pillow
anthropic
fish-audio-sdk
